requirements.txt
# Core dependencies
datasets>=2.14.0
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# API clients
openai>=1.0.0
anthropic>=0.7.0
requests>=2.28.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.64.0
jinja2>=3.1.0

# Server management
psutil>=5.9.0

# Optional dependencies for cloud providers
boto3>=1.26.0  # For AWS Bedrock
google-cloud-aiplatform>=1.25.0  # For Google Cloud

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Jupyter for experimentation
jupyter>=1.0.0
ipywidgets>=8.0.0

# Visualization (optional)
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0
